Resolve log folder not being created in test
There is presently an error when running make && make test where
there is no log folder being created in the test application. This
is likely due to there being no Lager command run, which would
otherwise generate the folder. There is one being generated in
the root folder, but that is based on lager's source code.

Perhaps there is a more elegant way to do this or the folder does
not need to be generated at all, but for the time being does create
it and resolves the test failure.